+++ src/pyird/data/channel_H.list
@@ -1,0 +1,69 @@
+ORDER,CHANNEL,WAVELENGTH,ST
+1,905,1417.25,0
+1,970,1417.86,0
+1,1981,1425.31,0
+2,370,1425.31,0
+2,437,1426.08,0
+2,414,1425.80,1
+3,689,1442.85,0
+3,960,1445.57,1
+4,976,1460.05,0
+4,1412,1463.84,0
+4,1627,1465.44,0
+4,968,1459.97,1
+4,1386,1463.64,1
+5,18,1463.64,1
+5,33,1463.84,0
+5,156,1465.44,0
+5,192,1465.89,1
+5,508,1469.70,0
+5,942,1474.32,0
+5,1480,1479.01,0
+6,61,1479.01,0
+6,827,1488.06,0
+6,1547,1494.46,0
+6,1595,1494.81,0
+7,1007,1505.06,0
+7,848,1503.46,0
+7,854,1503.53,0
+8,728,1517.68,0
+8,774,1518.19,1
+9,478,1530.61,0
+9,714,1533.35,0
+9,901,1535.35,0
+9,939,1535.73,0
+9,1486,1540.69,0
+9,1868,1543.40,0
+10,14,1540.69,0
+10,214,1543.40,0
+10,346,1545.10,0
+10,954,1552.07,0
+10,1372,1555.97,0
+10,1640,1558.08,0
+11,85,1558.08,1
+11,1542,1573.92,0
+12,19,1573.92,1
+12,652,1582.11,0
+12,783,1583.61,0
+12,1294,1588.75,0
+12,1487,1590.40,0
+13,658,1599.39,0
+14,319,1612.71,0
+14,784,1618.44,0
+14,1677,1626.85,0
+15,47,1626.85,0
+15,1572,1644.11,0
+16,585,1652.44,0
+16,832,1655.38,0
+16,842,1655.49,0
+17,856,1674.46,0
+17,1119,1677.31,0
+18,926,1694.52,0
+19,1661,1721.29,0
+20,573,1730.08,0
+20,663,1731.24,0
+20,1319,1738.53,0
+20,1333,1738.67,0
+20,1555,1740.67,0
+21,164,1744.97,0
+21,1289,1758.93,0
