+++ src/pyird/data/channel_YJ.list
@@ -1,0 +1,187 @@
+ORDER,CHANNEL,WAVELENGTH,ST
+1,1423,935.68,0
+2,459,935.68,0
+2,1147,940.17,0
+3,1436,947.7478,0
+3,1903,949.8102,1
+3,1949,949.9794,1
+4,749,949.8102,0
+4,774,949.9794,0
+6,1379,966.0435,0
+7,404,966.0435,0
+7,1018,970.32230,0
+-7,0,970.32712,0
+7,1929,974.9135,0
+8,740,974.9135,1
+8,1356,978.7186,0
+8,1972,981.5388,0
+9,378,978.7186,0
+9,757,981.5388,0
+9,963,982.9144,1
+9,1073,983.6119,1
+11,1444,999.0373,0
+12,1158,1004.2116,0
+12,1378,1005.4796,0
+13,865,1009.190,0
+13,1214,1011.4365,1
+13,1622,1013.6343,0
+13,1792,1014.4176,0
+16,1367,1033.5556,0
+17,1630,1042.2433,0
+18,1165,1047.0049,1
+18,1211,1047.2922,0
+18,1344,1048.0904,0
+18,1913,1050.9380,1
+19,224,1047.2922,0
+19,315,1048.0904,0
+19,662,1050.9380,0
+19,1405,1055.9342,0
+19,1569,1056.8199,0
+21,856,1067.6489,0
+21,971,1068.4699,0
+21,994,1068.6331,0
+21,1263,1070.3918,0
+21,1735,1072.9864,0
+22,542,1072.9864,0
+22,627,1073.6795,1
+22,962,1076.2093,0
+22,1172,1077.6320,0
+23,1659,1088.3923,0
+23,1306,1086.4053,1
+24,484,1088.3923,0
+24,1317,1094.5242,0
+24,1460,1095.3727,0
+25,354,1095.3727,0
+25,1832,1105.4921,0
+26,569,1105.4921,1
+26,909,1108.1903,0
+26,1311,1110.9492,0
+28,683,1123.3329,0
+28,982,1125.6586,0
+29,1146,1135.5028,0
+29,1187,1135.7822,0
+29,1866,1139.6820,0
+30,558,1139.6820,0
+30,1004,1143.2836,0
+30,1174,1144.4964,0
+30,1585,1147.0685,0
+30,1996,1149.1254,0
+31,125,1144.4964,0
+31,391,1147.0685,0
+31,620,1149.1254,0
+31,906,1151.4709,0
+32,1990,1167.1904,0
+33,600,1167.1904,0
+33,713,1168.1683,0
+33,1024,1170.6656,0
+33,1248,1172.2696,0
+33,1460,1173.6442,0
+34,154,1172.2696,0
+34,291,1173.6442,0
+34,1125,1180.7847,0
+35,682,1186.7493,0
+35,869,1188.3237,0
+35,924,1188.7733,0
+35,1082,1189.9889,0
+35,1776,1194.3905,0
+35,1828,1194.6547,0
+35,1840,1194.7135,0
+36,462,1194.3905,0
+36,490,1194.6547,0
+36,496,1194.7135,0
+36,969,1198.7948,0
+36,1448,1202.2005,0
+36,1577,1202.9940,0
+37,340,1202.9940,0
+37,1054,1209.2871,1
+37,1372,1211.5641,0
+37,1612,1213.0619,1
+37,1839,1214.3061,0
+38,205,1211.5641,0
+38,352,1213.0619,1
+38,479,1214.3061,0
+38,603,1215.4691,0
+38,1115,1219.7499,1
+38,1290,1221.0233,0
+38,1688,1223.5291,0
+38,1711,1223.6544,1
+38,1731,1223.7723,1
+39,388,1223.5291,0
+39,1761,1234.1373,0
+39,1864,1234.6771,0
+40,475,1234.6771,0
+40,610,1235.9673,0
+40,1163,1240.6221,0
+40,1395,1242.2787,0
+40,1719,1244.2726,0
+41,45,1240.6221,0
+41,197,1242.2787,1
+41,388,1244.2726,0
+41,559,1245.9522,0
+41,789,1248.0709,1
+41,909,1249.1079,0
+41,1897,1255.7762,1
+42,479,1255.7762,0
+42,934,1259.9655,0
+42,1253,1262.5071,0
+42,1497,1264.1939,0
+42,1625,1264.9994,0
+43,89,1262.5071,0
+43,243,1264.1939,0
+43,318,1264.9994,0
+43,900,1270.5757,0
+43,1293,1273.6901,0
+43,1478,1274.9718,0
+44,106,1273.6901,0
+44,222,1274.9718,0
+44,575,1278.6274,1
+44,787,1280.6241,0
+44,909,1281.6987,1
+44,1627,1287.0160,0
+45,993,1293.6732,0
+45,1084,1294.4203,1
+45,1228,1295.5625,1
+45,1291,1296.0204,0
+45,1333,1296.3362,1
+45,1783,1299.2253,1
+46,88,1296.0204,0
+46,377,1299.2253,1
+46,565,1301.1821,0
+46,752,1302.9748,1
+46,777,1303.1989,0
+46,1035,1305.4747,1
+46,1091,1305.9458,1
+46,1284,1307.4405,0
+47,788,1314.9507,1
+47,875,1315.7435,1
+47,1532,1320.8895,1
+47,1669,1321.76071,0
+47,1916,1323.1723,0
+47,1971,1323.4513,0
+48,296,1321.76071,0
+48,426,1323.1723,0
+48,453,1323.4513,0
+48,878,1327.6266,0
+48,1230,1330.5951,0
+48,1375,1331.6851,0
+48,1623,1333.3763,0
+49,113,1331.6851,0
+49,260,1333.3763,0
+49,606,1337.0767,0
+49,613,1337.1482,1
+49,1027,1341.0256,0
+50,241,1345.4007,1
+50,443,1347.6409,1
+50,701,1350.3096,0
+50,751,1350.7883,0
+50,899,1352.1843,1
+50,1202,1354.7905,0
+50,1487,1356.9371,0
+50,1604,1357.7328,0
+51,162,1356.9371,0
+51,230,1357.7328,0
+51,460,1360.3052,0
+51,684,1362.6384,0
+51,795,1363.7282,1
+51,1315,1368.2290,1
+
