+++ src/pyird/image/readnoise.py
