--- src/pyird/pyird_version.py
@@ -1,1 +1,0 @@
-__version__ = "0.1.dev32+gbf317f6"
