+++ src/pyird/utils/fitsset.py
