+++ src/pyird/utils/getinfo.py
