+++ src/pyird/utils/irdstream.py
