+++ src/pyird/utils/makemask.py
