+++ src/pyird/utils/pyird_version.py
@@ -1,0 +1,1 @@
+__version__ = "0.0.2.dev14+g92c5019.d20211213"
