+++ tests/utils/test_bias.py
@@ -1,0 +1,8 @@
+import pytest
+from pyird.image import bias
+
+def test_rmbias():
+    assert False
+
+if __name__=="__main__":
+    test_rmbias()
